POWER_HAL: Fix code for 64 bits
One minor fix of a cast from a pointer to an int. Such direct
casts are not possible when pointers are 64 bits and integers
are still 32 bits.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-9533
Reviewed-on: https://android.intel.com:443/299921

+++ power_intel.c
@@ -288,7 +288,7 @@ static void intel_power_hint(struct power_module *module, power_hint_t hint,
                 }
             }
             if (intel->vsync_boost) {
-                if (((int)data != 0) && (vsync_count > 0)) {
+                if (((unsigned long)data != 0) && (vsync_count > 0)) {
                     sysfs_write(TOUCHBOOST_PULSE_SYSFS,"1");
                     vsync_count-- ;
                 if (vsync_count == 0)
